fix: ci build
